Module bevy_crevice::std140

Defines traits and types for working with data adhering to GLSL’s std140 layout specification.

Structs

DMat2

Corresponds to a GLSL dmat2 in std140 layout.

DMat3

Corresponds to a GLSL dmat3 in std140 layout.

DMat4

Corresponds to a GLSL dmat3 in std140 layout.

DVec2

Corresponds to a GLSL dvec2 in std140 layout.

DVec3

Corresponds to a GLSL dvec3 in std140 layout.

DVec4

Corresponds to a GLSL dvec4 in std140 layout.

DynamicUniform

Wrapper type that aligns the inner type to at least 256 bytes.

DynamicUniformStd140

std140 variant of DynamicUniform.

IVec2

Corresponds to a GLSL ivec2 in std140 layout.

IVec3

Corresponds to a GLSL ivec3 in std140 layout.

IVec4

Corresponds to a GLSL ivec4 in std140 layout.

Mat2

Corresponds to a GLSL mat2 in std140 layout.

Mat3

Corresponds to a GLSL mat3 in std140 layout.

Mat4

Corresponds to a GLSL mat4 in std140 layout.

Sizer

Type that computes the buffer size needed by a series of std140 types laid out.

UVec2

Corresponds to a GLSL uvec2 in std140 layout.

UVec3

Corresponds to a GLSL uvec3 in std140 layout.

UVec4

Corresponds to a GLSL uvec4 in std140 layout.

Vec2

Corresponds to a GLSL vec2 in std140 layout.

Vec3

Corresponds to a GLSL vec3 in std140 layout.

Vec4

Corresponds to a GLSL vec4 in std140 layout.

Writer

Type that enables writing correctly aligned std140 values to a buffer.

Enums

InvalidPadded

Unfortunately, we cannot easily derive padded representation for generic Std140 types. For now, we’ll just use this empty enum with no values.

Traits

AsStd140

Trait implemented for all types that can be turned into std140 values.

Std140

Trait implemented for all std140 primitives. Generally should not be implemented outside this crate.

Std140Convertible

Trait specifically for Std140::Padded, implements conversions between padded type and base type.

WriteStd140

Trait implemented for all types that can be written into a buffer as std140 bytes. This type is more general than AsStd140: all AsStd140 types implement WriteStd140, but not the other way around.

Derive Macros

AsStd140